# api/swagger.yaml
definitions:
...
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
  controller.bindEmailRequest:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  controller.sendEmailRequest:
    properties:
      capt:
        type: string
      email:
        type: string
    required:
    - capt
    - email
    type: object
...
  controller.tokenCreateRequest:
    properties:
      name:
        type: string
      perm:
        type: string
    required:
    - name
    - perm
    type: object
  controller.userBasicInfoUpdateRequest:
    properties:
      avatar_id:
        type: string
      description:
        type: string
      email:
        type: string
      fullname:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/user:
    get:
      consumes:
      - application/json
      description: get current sign-in user info
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: Get current user info
      tags:
      - User
    put:
      consumes:
      - application/json
      description: update user basic info
      parameters:
      - description: body of updating user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.userBasicInfoUpdateRequest'
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: Update
      tags:
      - User
  /v1/user/email/bind:
    post:
      consumes:
      - application/json
      description: bind user's email
      parameters:
      - description: body of bind email info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.bindEmailRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: Bind User Email
      tags:
      - User
  /v1/user/email/send:
    post:
      consumes:
      - application/json
      description: send user's email verify code
      parameters:
      - description: body of bind email info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.sendEmailRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: Send User Email Verify code
      tags:
      - User
  /v1/user/token:
    get:
      consumes:
      - application/json
      description: list all platform tokens of user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: GetTokenInfo
      tags:
      - User
    post:
      consumes:
      - application/json
      description: create a new platform token of user
      parameters:
      - description: body of create token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.tokenCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: CreatePlatformToken
      tags:
      - User
  /v1/user/token/{name}:
    delete:
      consumes:
      - application/json
      description: delete a new platform token of user
      parameters:
      - description: token name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: DeletePlatformToken
      tags:
      - User
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and api Bearer.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"